feat(moss-cache): Basic Cache API and tests
