--- crates/moss-store/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "moss-store"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-moka = {version = "0.12.8", features = ["sync"]}
-dashmap = "6.1.0"
