--- crates/moss-store/src/lib.rs
@@ -1,1 +1,0 @@
-mod moss_cache;
